requirements.txt
botbuilder-integration-aiohttp>=4.16.0
aiohttp>=3.9.0
msal>=1.28.0
anthropic>=0.42.0
python-dotenv>=1.0.0
